requirements (1).txt
streamlit>=1.33
Pillow>=10.0
python-pptx>=0.6.22
